requirements.txt
Django
djangorestframework
pylint-django
djangorestframework-simplejwt
django-cors-headers
psycopg2
django-storages[azure]
urllib3==1.25.3
